FIX: make Dinode private and revert 0238372
